+++ NEWS.md
@@ -1,9 +1,10 @@
 
-# rquery 1.3.9 2019/09/29
+# rquery 1.3.9 2019/10/03
 
  * Tune yaml path.
  * Fold consecutive select_columns nodes.
  * Don't print quoted table name.
+ * Add arrow example.
 
 # rquery 1.3.8 2019/09/15